[TensorExpr] Add methods for inspecting generated code in `TensorExprKernel`.
The new methods allow to peak into bufferArgs which describe parameters
that codegen expects. This description includes info whether a given
parameter is a scalar var or a buffer and in case it's a buffer allows
to get the corresponding `Buf*` pointer from which we could get the
expected sizes.

ghstack-source-id: 20c67a82a96112a13ca21fa2e6faac65bcc66f19
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57074