[TensorExpr] Remove `dtype_` and add `buf_` fields to `CodeGen::BufferArg`.
`BufferArg` is used to describe parameters passed to the codegen: it
indicates whether the parameter is a var or a buf and holds a pointer to
the corresponding var/buf. Both var and buf contain dtype, and thus
duplicating it in BufferArg is unnecessary - we can always get it from
the var/buf. Hence we're removing dtype_ field from BufferArg in this
PR. We're also adding a `buf_` field here: this is done so that
BufferArg truly has all the info about the parameter.

ghstack-source-id: 114f9b9e4acdedb18fd2e17dc7ef920ace7a7b9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57382