[RFC] [PyTorch Edge] Simplify error logging in mobile/import.cpp
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55711

Currently, there is some complex logic that tries to handle all exceptions but re-throws them as a `c10::Error` so that it can log the error message. I'm looking for context on why this was added. The current logic (after talking with @swolchok) seems equivalent, simpler, and also preserves the original stack trace from where the exception was originally thrown. This is useful when viewing the backtrace in logview. Re-throwing an exception using `TORCH_CHECK(false, message)` results in the original exception stack trace getting lost, so we want to avoid that.
ghstack-source-id: 128043281

Differential Revision: [D27688352](https://our.internmc.facebook.com/intern/diff/D27688352/)