[package] fix corner case in PacakgeImporter.whichmodule
We've gone back and forth on whether to emulate the `sys.modules` lookup
behavior in our own `whichmodule`, the provided test is a concrete case
for doing so.

An additional minor cleanup is to make the type of `self.modules` in
importers `Dict[str, ModuleType]`. Modules could only be None in the
dictionary in older versions of the import system

ghstack-source-id: 2bdc45bee8814ceea18e47df70eca15b7575fc11
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57651